+++ pufferlib/ocean/drone_fork/binding.c
@@ -1,4 +1,4 @@
-#include "drone_swarm.h"
+#include "drone_fork.h"
 
 #define Env DroneSwarm
 #include "../env_binding.h"
@@ -13,9 +13,9 @@ static int my_init(Env *env, PyObject *args, PyObject *kwargs) {
 static int my_log(PyObject *dict, Log *log) {
     assign_to_dict(dict, "perf", log->perf);
     assign_to_dict(dict, "score", log->score);
-    assign_to_dict(dict, "rings_passed", log->rings_passed);
     assign_to_dict(dict, "collision_rate", log->collision_rate);
     assign_to_dict(dict, "oob", log->oob);
+    assign_to_dict(dict, "rocket_hits", log->rocket_hits);
     assign_to_dict(dict, "episode_return", log->episode_return);
     assign_to_dict(dict, "episode_length", log->episode_length);
     assign_to_dict(dict, "n", log->n);
